Prevent failures for concurrent update_helm_index workflows (#35)
The solution is to request exclusive access BEFORE checking out the
website repository. We also ensure that the latest version of main is
checked out, not the commit for which the workflow was triggered.

As part of this change, we update the EndBug/add-and-commit action to
the latest version. One important change is that starting with v8, the
action will now work with whatever ref has been checked out, without
pulling or switching branches by default.

Fixes #34

Signed-off-by: Antonin Bas <abas@vmware.com>